add console output as setting

+++ core/backlight.c
@@ -57,7 +57,7 @@ static void backlight_write(const uint16_t pio, const uint8_t byte, bool poke) {
             break;
     }
 
-    backlight.factor = (310 - (float)backlight.brightness) / 160.0;
+    backlight.factor = (310 - (float)backlight.brightness) / (float)160.0;
 }
 
 static const eZ80portrange_t device = {
